refactor: remove Card wrapper and standardize layout across auth and connection components
