+++ src/components/Connections/ConnectSmartcar.tsx
@@ -1,10 +1,8 @@
 import { type FC } from 'react';
 
 import VehicleThumbnail from '../../assets/images/vehicle-thumbnail.png';
-import PrimaryButton from '../Shared/PrimaryButton';
+import { PrimaryButton, Header } from '../Shared';
 import { UiStates, useUIManager } from '../../context/UIManagerContext';
-import Card from '../Shared/Card';
-import Header from '../Shared/Header';
 import { getAppUrl } from '../../utils/urlHelpers';
 import { useDevCredentials } from '../../context/DevCredentialsContext';
 import { constructAuthUrl } from '../../utils/authUrls';
@@ -36,37 +34,31 @@ export const ConnectSmartCar: FC = () => {
   };
 
   return (
-    <Card
-      width="w-full max-w-[600px]"
-      height="h-fit"
-      className="flex flex-col items-center"
-    >
-      <div className="flex flex-col gap-6 w-[440px]">
-        <Header
-          title={`Connect Data Source for ${componentData.modelYear} ${componentData.makeModel}`}
-          subtitle={appUrl.hostname}
-        />
+    <>
+      <Header
+        title={`Connect Data Source for ${componentData.modelYear} ${componentData.makeModel}`}
+        subtitle={appUrl.hostname}
+      />
 
-        <div className="w-full text-gray-600 text-sm text-center">
-          At least one data source is required to connect to {devLicenseAlias}. Connecting
-          through your car's app is the easiest option and will take less than a minute.
-        </div>
-
-        <div className="flex justify-center pt-2">
-          <img
-            style={{ height: '120px', width: '120px' }}
-            className="rounded-full object-cover"
-            src={VehicleThumbnail}
-            alt="DIMO Adapter"
-          />
-        </div>
-
-        <div className="flex flex-col w-full space-y-3">
-          <PrimaryButton onClick={handleConnect} width="w-full py-3">
-            Connect with Smartcar
-          </PrimaryButton>
-        </div>
+      <div className="w-full text-gray-600 text-sm text-center">
+        At least one data source is required to connect to {devLicenseAlias}. Connecting
+        through your car's app is the easiest option and will take less than a minute.
       </div>
-    </Card>
+
+      <div className="flex justify-center pt-2">
+        <img
+          style={{ height: '120px', width: '120px' }}
+          className="rounded-full object-cover"
+          src={VehicleThumbnail}
+          alt="DIMO Adapter"
+        />
+      </div>
+
+      <div className="flex flex-col w-full mt-3">
+        <PrimaryButton onClick={handleConnect} width="w-full py-3">
+          Connect with Smartcar
+        </PrimaryButton>
+      </div>
+    </>
   );
 };
